+++ blockstack-sdk/src/main/java/org/blockstack/android/sdk/SessionStore.kt
@@ -1,5 +1,6 @@
 package org.blockstack.android.sdk
 
+import android.content.Context
 import android.content.SharedPreferences
 import android.util.Log
 import org.blockstack.android.sdk.model.SessionData
@@ -46,5 +47,7 @@ class SessionStore(private val prefs: SharedPreferences) : ISessionStore {
         sessionDataObject.json.put("userData", userData.json)
         prefs.edit().putString(BLOCKSTACK_SESSION, sessionDataObject.json.toString()).apply()
     }
-
 }
+
+fun Context.getBlockstackSharedPreferences(): SharedPreferences =
+        getSharedPreferences("${packageName}_blockstack_prefs", Context.MODE_PRIVATE)