+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - additional parameters `sendToSignIn`, `appDetails` for `BlockstackSignIn`
 - `ConnectBottemSheetDialogFragment` and `SignInProvider` to show user on-boarding
+- `Context.getBlockstackSharedPreferences` to be used for `SessionStore`
+- `AppCompatActivity.showBlockstackConnect` for simplified integration
 ### Changed
 - use app.blockstack.org (connect) instead of browser.blockstack.org for sign in.